Improved address input appearance and validation, moved JS to separate file
